Re-factored persistence layer to be less monolithic
