+++ src/Altinn.Broker.Application/InitializeFileCommand/InitializeFileCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Altinn.Broker.Core.Application;
+using Altinn.Broker.Core.Domain.Enums;
 using Altinn.Broker.Core.Repositories;
 using Altinn.Broker.Core.Services;
 
@@ -11,13 +12,17 @@ public class InitializeFileCommandHandler : IHandler<InitializeFileCommandReques
 {
     private readonly IServiceOwnerRepository _serviceOwnerRepository;
     private readonly IFileRepository _fileRepository;
+    private readonly IFileStatusRepository _fileStatusRepository;
+    private readonly IActorFileStatusRepository _actorFileStatusRepository;
     private readonly IResourceManager _resourceManager;
     private readonly ILogger<InitializeFileCommandHandler> _logger;
 
-    public InitializeFileCommandHandler(IServiceOwnerRepository serviceOwnerRepository, IFileRepository fileRepository, IResourceManager resourceManager, ILogger<InitializeFileCommandHandler> logger)
+    public InitializeFileCommandHandler(IServiceOwnerRepository serviceOwnerRepository, IFileRepository fileRepository, IFileStatusRepository fileStatusRepository, IActorFileStatusRepository actorFileStatusRepository, IResourceManager resourceManager, ILogger<InitializeFileCommandHandler> logger)
     {
         _serviceOwnerRepository = serviceOwnerRepository;
         _fileRepository = fileRepository;
+        _fileStatusRepository = fileStatusRepository;
+        _actorFileStatusRepository = actorFileStatusRepository;
         _resourceManager = resourceManager;
         _logger = logger;
     }
@@ -39,6 +44,16 @@ public class InitializeFileCommandHandler : IHandler<InitializeFileCommandReques
             return Errors.ServiceOwnerNotReadyInfrastructure;
         }
         var fileId = await _fileRepository.AddFile(serviceOwner, request.Filename, request.SendersFileReference, request.SenderExternalId, request.RecipientExternalIds, request.PropertyList, request.Checksum);
+        await _fileStatusRepository.InsertFileStatus(fileId, FileStatus.Initialized);
+        var addRecipientEventTasks = request.RecipientExternalIds.Concat([request.SenderExternalId]).Select(recipientId => _actorFileStatusRepository.InsertActorFileStatus(fileId, ActorFileStatus.Initialized, recipientId));
+        try
+        {
+            await Task.WhenAll(addRecipientEventTasks);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError("Failed when adding recipient initialized events.");
+        }
         return fileId;
     }
 }
