+++ src/Altinn.Broker.Core/Repositories/IFileRepository.cs
@@ -15,17 +15,9 @@ public interface IFileRepository
         string? checksum);
     Task<Domain.FileEntity?> GetFile(Guid fileId);
     Task<List<Guid>> GetFilesAssociatedWithActor(ActorEntity actor);
-    Task AddReceipt(
-        Guid fileId,
-        Domain.Enums.ActorFileStatus status,
-        string actorExternalReference
-    );
     Task SetStorageReference(
         Guid fileId,
         long storageProviderId,
         string fileLocation
     );
-    Task<List<FileStatusEntity>> GetFileStatusHistory(Guid fileId);
-    Task<List<ActorFileStatusEntity>> GetActorEvents(Guid fileId);
-    Task InsertFileStatus(Guid fileId, FileStatus status);
 }
