+++ src/Altinn.Broker.Persistence/Repositories/FileRepository.cs
@@ -1,6 +1,4 @@
-﻿using System.Data;
-
-using Altinn.Broker.Core.Domain;
+﻿using Altinn.Broker.Core.Domain;
 using Altinn.Broker.Core.Domain.Enums;
 using Altinn.Broker.Core.Repositories;
 
@@ -134,19 +132,7 @@ public class FileRepository : IFileRepository
 
         command.ExecuteNonQuery();
 
-        var addRecipientEventTasks = recipientIds.Concat([senderExternalId]).Select(recipientId => AddReceipt(fileId, ActorFileStatus.Initialized, recipientId));
-
-        try
-        {
-            await Task.WhenAll(addRecipientEventTasks);
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"An error occurred: {ex.Message}");
-        }
-
         await SetMetadata(fileId, propertyList);
-        await InsertFileStatus(fileId, FileStatus.Initialized);
 
         return fileId;
     }
@@ -185,36 +171,6 @@ public class FileRepository : IFileRepository
         }
     }
 
-    public async Task AddReceipt(Guid fileId, ActorFileStatus status, string actorExternalReference)
-    {
-        var connection = await _connectionProvider.GetConnectionAsync();
-
-        var actor = await _actorRepository.GetActorAsync(actorExternalReference);
-        long actorId = 0;
-        if (actor is null)
-        {
-            actorId = await _actorRepository.AddActorAsync(new ActorEntity()
-            {
-                ActorExternalId = actorExternalReference
-            });
-        }
-        else
-        {
-            actorId = actor.ActorId;
-        }
-
-        using (var command = new NpgsqlCommand(
-            "INSERT INTO broker.actor_file_status (actor_id_fk, file_id_fk, actor_file_status_id_fk, actor_file_status_date) " +
-            "VALUES (@actorId, @fileId, @actorFileStatusId, NOW())", connection))
-        {
-            command.Parameters.AddWithValue("@actorId", actorId);
-            command.Parameters.AddWithValue("@fileId", fileId);
-            command.Parameters.AddWithValue("@actorFileStatusId", (int)status);
-            var commandText = command.CommandText;
-            command.ExecuteNonQuery();
-        }
-    }
-
     public async Task SetStorageReference(Guid fileId, long storageProviderId, string fileLocation)
     {
         var connection = await _connectionProvider.GetConnectionAsync();
@@ -230,66 +186,6 @@ public class FileRepository : IFileRepository
             command.Parameters.AddWithValue("@storageProviderId", storageProviderId);
             command.Parameters.AddWithValue("@fileLocation", fileLocation);
             command.ExecuteNonQuery();
-        }
-    }
-
-    public async Task<List<FileStatusEntity>> GetFileStatusHistory(Guid fileId)
-    {
-        var connection = await _connectionProvider.GetConnectionAsync();
-
-        using (var command = new NpgsqlCommand(
-            "SELECT * " +
-            "FROM broker.file_status fis " +
-            "WHERE fis.file_id_fk = @fileId", connection))
-        {
-            command.Parameters.AddWithValue("@fileId", fileId);
-            var fileStatuses = new List<FileStatusEntity>();
-            using (var reader = await command.ExecuteReaderAsync())
-            {
-                while (await reader.ReadAsync())
-                {
-                    fileStatuses.Add(new FileStatusEntity()
-                    {
-                        FileId = reader.GetGuid(reader.GetOrdinal("file_id_fk")),
-                        Status = (FileStatus)reader.GetInt32(reader.GetOrdinal("file_status_description_id_fk")),
-                        Date = reader.GetDateTime(reader.GetOrdinal("file_status_date")),
-                    });
-                }
-            }
-            return fileStatuses;
-        }
-    }
-
-    public async Task<List<ActorFileStatusEntity>> GetActorEvents(Guid fileId)
-    {
-        var connection = await _connectionProvider.GetConnectionAsync();
-
-        using (var command = new NpgsqlCommand(
-            "SELECT *, a.actor_external_id " +
-            "FROM broker.actor_file_status afs " +
-            "INNER JOIN broker.actor a on a.actor_id_pk = afs.actor_id_fk " +
-            "WHERE afs.file_id_fk = @fileId", connection))
-        {
-            command.Parameters.AddWithValue("@fileId", fileId);
-            var fileStatuses = new List<ActorFileStatusEntity>();
-            using (var reader = await command.ExecuteReaderAsync())
-            {
-                while (await reader.ReadAsync())
-                {
-                    fileStatuses.Add(new Core.Domain.ActorFileStatusEntity()
-                    {
-                        FileId = reader.GetGuid(reader.GetOrdinal("file_id_fk")),
-                        Status = (Core.Domain.Enums.ActorFileStatus)reader.GetInt32(reader.GetOrdinal("actor_file_status_id_fk")),
-                        Date = reader.GetDateTime(reader.GetOrdinal("actor_file_status_date")),
-                        Actor = new ActorEntity()
-                        {
-                            ActorId = reader.GetInt64(reader.GetOrdinal("actor_id_fk")),
-                            ActorExternalId = reader.GetString(reader.GetOrdinal("actor_external_id"))
-                        }
-                    });
-                }
-            }
-            return fileStatuses;
         }
     }
 
@@ -345,24 +241,6 @@ public class FileRepository : IFileRepository
         {
             transaction.Rollback();
             throw;
-        }
-    }
-
-    public async Task InsertFileStatus(Guid fileId, FileStatus status)
-    {
-        var connection = await _connectionProvider.GetConnectionAsync();
-        using var command = new NpgsqlCommand("", connection);
-
-        command.CommandText =
-            "INSERT INTO broker.file_status (file_id_fk, file_status_description_id_fk, file_status_date) " +
-            "VALUES (@fileId, @statusId, NOW()) RETURNING file_status_id_pk;";
-        command.Parameters.AddWithValue("@fileId", fileId);
-        command.Parameters.AddWithValue("@statusId", (int)status);
-
-        var fileStatusId = await command.ExecuteScalarAsync();
-        if (fileStatusId == null)
-        {
-            throw new InvalidOperationException("No file_status_id_pk was returned after insert.");
         }
     }
 }
